requirements.txt
bcrypt
email-validator
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
pep8
psycopg2
python-dotenv
SQLAlchemy
WTForms
Flask-Caching
redis
gunicorn
